Small changes in models (decimal precision).

+++ DAL/DbModel/Models/FoodMixtureItem.cs
@@ -5,6 +5,6 @@
 		public int Id { get; set; }
 		public FoodMixture FoodMixture { get; set; }
 		public FoodProductItem FoodProductItem { get; set; }
-		public decimal ItemWeightInGram { get; set; }
+		public int ItemWeightInGram { get; set; }
 	}
 }
